requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate
Flask-WTF
Flask-Cors
python-dotenv==1.0.1
APScheduler
twilio
Pillow
requests
boto3
botocore
sendgrid
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
psycopg2==2.9.10
SQLAlchemy==2.0.37
Werkzeug==3.1.3
